requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
gevent==1.2.2
greenlet==0.4.12
idna==2.6
Pillow==5.0.0
python-dotenv==0.7.1
qrcode==5.3
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
